fix: remove google-services.json from git tracking and update .gitignore
- Add google-services.json to .gitignore (contains Firebase API keys)
- Remove tracked google-services.json from index (file stays on disk)
- Add gitignore entries for local.properties, keystores, APKs, heap dumps

Closes #14
