--- Final Submission Project/app/google-services.json
@@ -1,29 +1,0 @@
-{
-  "project_info": {
-    "project_number": "123456789012",
-    "project_id": "audiobook-player-dev",
-    "storage_bucket": "audiobook-player-dev.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:123456789012:android:abcdef123456789",
-        "android_client_info": {
-          "package_name": "com.audiobook.app"
-        }
-      },
-      "oauth_client": [],
-      "api_key": [
-        {
-          "current_key": "AIzaSyDummyKeyForDevelopmentPurposesOnly"
-        }
-      ],
-      "services": {
-        "appinvite_service": {
-          "other_platform_oauth_client": []
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
