--- requirements.txt
@@ -1,3 +1,0 @@
-mutagen>=1.47.0
-requests>=2.31.0
-colorama>=0.4.6
